Refactor: Split drawflow.js into modular files with different responsibilities
